refs #54319 @40m вычистил лишний код
